+++ app/routes/question.js
@@ -6,13 +6,15 @@ export default Ember.Route.extend({
   },
   actions: {
     deleteThisQuestion(question) {
-      var answer_deletions = question.get('answers').map(function(answer) {
-        return answer.destroyRecord();
-      });
-      Ember.RSVP.all(answer_deletions).then(function() { //Ember.RSVP.all() can package many promises and wait for them.
-        return question.destroyRecord(); //promised callback function that deletes from Firebase
-      });
-      this.transitionTo('index');
+      if(confirm('All answers belonging to it will be lost as well. Are you sure you want to delete this question?')) {
+        var answer_deletions = question.get('answers').map(function(answer) {
+          return answer.destroyRecord();
+        });
+        Ember.RSVP.all(answer_deletions).then(function() { //Ember.RSVP.all() can package many promises and wait for them.
+          return question.destroyRecord(); //promised callback function that deletes from Firebase
+        });
+        this.transitionTo('index');
+      }
     },
     updateQuestion(question, params) {
       Object.keys(params).forEach(function(key) {
